refactor(types): unify duplicate SlotRange, NodeRole, and OperationType
Three SlotRange structs collapse into one canonical type at
slots::distribution::SlotRange with u16 fields (the correct domain
type — slots are 0..=16383). Removed duplicates from client::types
and controller::cluster_topology, which now re-export the canonical
type. Added SlotRange::new, ::single, ::parse_valkey, ::count,
::contains, ::iter helpers and the inevitable i32::from boundary
conversions where K8s/CRD types need i32.

Two NodeRole enums collapse into client::types::NodeRole with three
variants (Master, Replica, Unknown). Display divergence is now
explicit: as_valkey_str() for the wire format ("master"/"slave"/"?")
and as_display_str() for human text ("master"/"replica"/"unknown")
rather than hidden in two parallel Display impls. cluster_topology
re-exports for backward compat.

Two OperationType enums collapse into
controller::operation_coordination::OperationType, now Copy and
FromStr. operation_lock re-exports.

Removes 8 newly-redundant `as u16` casts and the i32 SlotRange
test helper signature.

+++ src/client/scaling.rs
@@ -249,7 +249,7 @@ impl ValkeyClient {
             );
             for range in &master.slots {
                 for slot in range.start..=range.end {
-                    current_ownership.insert(slot as u16, master.node_id.clone());
+                    current_ownership.insert(slot, master.node_id.clone());
                 }
             }
         }
@@ -452,7 +452,7 @@ impl ValkeyClient {
                     target
                         .slots
                         .iter()
-                        .any(|range| range.start <= slot as i32 && range.end >= slot as i32)
+                        .any(|range| range.start <= slot && range.end >= slot)
                 });
 
                 if owns_all_slots {
